refactor: LottoTicketValidation 명명 수정, private 생성자 생성
