+++ src/test/java/lotto/lottoticket/WinnerLottoTicketTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.Arrays;
 
 import static lotto.lottoticket.LottoNumber.ERROR_MESSAGE_INVALID_RANGE;
-import static lotto.lottoticket.TicketValidation.*;
+import static lotto.lottoticket.LottoTicketValidation.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
