+++ src/test/java/lotto/money/MoneyTest.java
@@ -3,7 +3,7 @@ package lotto.money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static lotto.lottoticket.TicketValidation.ERROR_MESSAGE_INVALID_INPUT;
+import static lotto.lottoticket.LottoTicketValidation.ERROR_MESSAGE_INVALID_INPUT;
 import static lotto.money.Money.ERROR_MESSAGE_MINIMUM_MONEY;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
